refactor(#47): 아이콘 네이밍 변경
