+++ src/components/mytrip/BookMarkBtn.tsx
@@ -1,30 +1,30 @@
 import { useState } from "react";
 
 import BookMarkIcon from "../../assets/icons/bookmark.svg?react";
-import FullBookMarkIcon from "../../assets/icons/fullbookmark.svg?react";
+import FullBookMarkIcon from "../../assets/icons/bookmark_filled_white.svg?react";
 
 interface Props {
-    currentBookMarked: boolean;
-    onClick: () => void;
+  currentBookMarked: boolean;
+  onClick: () => void;
 }
 
 function BookMarkBtn({ currentBookMarked, onClick }: Props) {
-    const [isBookMarked, setIsBookMarked] = useState<boolean>(currentBookMarked);
+  const [isBookMarked, setIsBookMarked] = useState<boolean>(currentBookMarked);
 
-    function onClickHandler() {
-        setIsBookMarked((prev) => !prev);
-        onClick();
-    }
+  function onClickHandler() {
+    setIsBookMarked((prev) => !prev);
+    onClick();
+  }
 
-    return (
-        <>
-            {isBookMarked ? (
-                <FullBookMarkIcon onClick={onClickHandler} />
-            ) : (
-                <BookMarkIcon onClick={onClickHandler} />
-            )}
-        </>
-    );
+  return (
+    <>
+      {isBookMarked ? (
+        <FullBookMarkIcon onClick={onClickHandler} />
+      ) : (
+        <BookMarkIcon onClick={onClickHandler} />
+      )}
+    </>
+  );
 }
 
 export default BookMarkBtn;
